README: Remove links to wrong build instructions in wiki
Someone relied on our documentation, and again
the docs turned out to contain wrong/outdated information:

https://mixxx.zulipchat.com/#narrow/stream/247620-development-help/topic/Build.20tools.20for.20Visual.20Studio.202022.3F/near/271368532

Documentation should be helpful and make things easier, but wrong
documentation is actually harmful.

This has not been the first time that our tendency to add redundant
doxumentation for ever edge case has bitten us.

Time and again it has been proven that we are not capable of keeping
these wordy guides up to date, so the links should be removed.